Complete tasks T49, T52, T53: module decomposition, prompt budget gate, command extraction
T49: Extract provenance bundle lifecycle, handoff note writing, and agent
identity tracking from iterationEngine.ts (1719→1301 lines) into
provenancePersistence.ts (474 lines).

T52: Add deterministic prompt budget validation gate using fixture-based
rendering (7 scenarios × 2 targets = 14 measurements) with committed
baseline and ±20% threshold.

T53: Decompose registerCommands.ts (1052→560 lines) by extracting 10
artifact-inspection and maintenance commands plus 7 helper functions
into artifactCommands.ts (430 lines).
